Create a partial basic implementation.
This partial implementation of the "basic hmplementation" only takes
one row or column of a board, and chooses the first valid word that
can be played, rather than the one that would give the highest score.
When given tile "u" and row "yo_", it correctly gives "you", but it
incorrectly considers a play invalid if the word would end before the
board row (eg with tiles "oul", row "c___d_", it does not find
"could_").