v1 basic: Validate words spelled indirectly.
When evaluating a possible play, look at the space around the proposed
word to check if it connects to any other words already on the board.
If so, consider the play valid only if these connections all produce
new, valid words.
Also replace output_n_times with a std::string constructor that
repeats chars similarly.