Prevent play of unconnected words.
Previously the human could play words disconnected from all other
words on the board (as long as they had the required letters), even
though this is considered invalid. Require that all words played be
connected to at least one previous word, with the exception of the
first word of the game.